Refactor collectors.System, changing metrics structure extensively.
Nest metrics structure. This is very much a backward-incompatible change,
but that's allowed at this point of the project. Also rename some metrics
slightly.

This very much reduces the size of the System payloads; in binary format,
it deducts around 26% of what it was previously. This both reduces data
over the wire and storage needed.